+++ example/TestLinkExample.py
@@ -88,6 +88,7 @@ NEWPROJECT="NEW_PROJECT_API-%i" % projNr
 NEWPREFIX="NPROAPI%i" % projNr
 NEWTESTPLAN_A="TestPlan_API A"
 NEWTESTPLAN_B="TestPlan_API B"
+NEWTESTPLAN_C="TestPlan_API C - DeleteTest"
 NEWPLATFORM_A='Big Birds %i' % projNr
 NEWPLATFORM_B='Small Birds'
 NEWPLATFORM_C='Ugly Birds'
@@ -99,6 +100,7 @@ NEWTESTCASE_B="TESTCASE_B"
 myApiVersion='%s v%s' % (myTestLink.__class__.__name__ , myTestLink.__version__)
 NEWBUILD_A='%s' % myApiVersion
 NEWBUILD_B='%s' % myApiVersion
+NEWBUILD_C='%s - DeleteTest' % myApiVersion
 
 this_file_dirname=os.path.dirname(__file__)
 NEWATTACHMENT_PY= os.path.join(this_file_dirname, 'TestLinkExample.py')
@@ -578,6 +580,47 @@ print("setTestCaseExecutionType", response)
 newResult = myTestLink.setTestCaseExecutionType(tc_b_full_ext_id, tc_version, 
                                                 newProjectID, MANUAL)
 print("setTestCaseExecutionType", response)
+
+# create TestPlan C with Platform, Build , TestCase, assigned TestCase 
+# and delete it 
+newTestPlan = myTestLink.createTestPlan(NEWTESTPLAN_C, NEWPROJECT,
+            notes='TestPlan for delete test.',
+            active=1, public=1)    
+print("createTestPlan for DeleteTest", newTestPlan)
+newTestPlanID_C = newTestPlan[0]['id'] 
+print("Test Plan '%s' - id: %s" % (NEWTESTPLAN_C,newTestPlanID_C))
+newBuild = myTestLink.createBuild(newTestPlanID_C, NEWBUILD_C, 
+                                  'Build for TestPlan delete test')
+print("createBuild for DeleteTest", newBuild)
+newBuildID_C = newBuild[0]['id'] 
+print("Build '%s' - id: %s" % (NEWBUILD_C, newBuildID_C))
+response = myTestLink.addPlatformToTestPlan(newTestPlanID_C, NEWPLATFORM_C) 
+print("addPlatformToTestPlan", response)
+response = myTestLink.addTestCaseToTestPlan(newProjectID, newTestPlanID_C, 
+                    tc_aa_full_ext_id, tc_version, platformid=newPlatFormID_C)
+print("addTestCaseToTestPlan", response)
+response = myTestLink.assignTestCaseExecutionTask( myTestUserName, 
+                        newTestPlanID_C, tc_aa_full_ext_id, buildid=newBuildID_C,
+                        platformid=newPlatFormID_C)  
+print("assignTestCaseExecutionTask", response)
+newResult = myTestLink.reportTCResult(newTestCaseID_AA, newTestPlanID_C, 
+                                      NEWBUILD_C, 'p', "TP delete test",
+                                      platformname=NEWPLATFORM_C)
+print("reportTCResult", newResult)
+newResultID_B = newResult[0]['id']
+newAttachment = myTestLink.uploadExecutionAttachment(NEWATTACHMENT_PY, newResultID_B, 
+        'Textfile Example', 'Attachment Example for a TC Execution and TP delete test',
+        filename='MyPyTPDeleteTest.py')
+print("uploadExecutionAttachment", newAttachment)
+response = myTestLink.getTotalsForTestPlan(newTestPlanID_C)
+print("getTotalsForTestPlan before delete", response)
+response = myTestLink.deleteTestPlan(newTestPlanID_C)
+print("deleteTestPlan", response)
+try:
+    response = myTestLink.getTotalsForTestPlan(newTestPlanID_C)
+    print("getTotalsForTestPlan after delete", response)
+except TLResponseError as tl_err:
+    print(tl_err.message)
 
 # get information - TestProject
 response = myTestLink.getTestProjectByName(NEWPROJECT)