+++ CHANGES.rst
@@ -46,6 +46,14 @@ new TestlinkAPIGeneric and TestlinkAPIClient api method
   
 examples see `<example/TestLinkExample_CF_KW.py>`_  
 
+implement 1.9.13 new api - deleteTestPlan #44 
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+new TestlinkAPIGeneric and TestlinkAPIClient api method
+
+- deleteTestPlan(<testplanid>)
+
+examples see `<example/TestLinkExample.py>`_ 
 
 TestLink-API-Python-client release notes v0.6.0 (Dec. 2014) 
 ------------------------------------------------------------